shelter: Move shader default source to renderer
This make sure that it is the renderer that controls what uniform should
be use.
